backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
geopy==2.3.0
firebase-admin==6.2.0
